mnist/requirements.txt
click==8.0.3
glob2==0.7
matplotlib==3.3.4
numpy==1.19.2
python-dotenv==0.19.2
setuptools==59.6.0
torch==1.10.1
